requirements.txt
streamlit>=1.32.0
torch>=2.2.0
torchvision>=0.17.0
transformers>=4.38.0
accelerate>=0.27.0
pillow>=10.2.0
pydantic>=2.6.0
numpy>=1.26.0
